requirements.txt
playwright
beautifulsoup4
requests
sqlite-utils
discord.py
aiohttp